# openapi.yaml
openapi: 3.0.3
info:
  title: DOGFOOD API
  version: 0.1.0
  description: 'Hackathon judging, voting, and management API. Note that admin/service routes typically wrap responses in { data: ... } while some public routes return unwrapped JSON. Error formats also vary between wrapped structured objects and raw string messages.'
servers:
  - url: 'http://localhost:3000'
    description: Local deployment
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorStandard:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/auth/register:
    post:
      summary: Register user
      description: Register a new account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - display_name
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                display_name:
                  type: string
                  minLength: 1
                  maxLength: 255
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStandard'
        '409':
          description: Duplicate email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStandard'
      security: []
      operationId: postapiauthregister
  /api/auth/login:
    post:
      summary: Login user
      description: Authenticate and return JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 1
                  maxLength: 128
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
                      expires_at:
                        type: string
                        format: date-time
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStandard'
      security: []
      operationId: postapiauthlogin
  /api/auth/logout:
    post:
      summary: Logout user
      description: Invalidate the current session token.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStandard'
      operationId: postapiauthlogout
  /api/auth/me:
    get:
      summary: Get current user
      description: Fetch the profile of the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStandard'
      operationId: getapiauthme
...